Don't import plain JS oblect, use deno-types pragma
