+++ basic.ts
@@ -1,8 +1,5 @@
-// Copyright 2020-present the denosaurs team. All rights reserved. MIT license.
-
-import type { Sodium } from "./basic_types.ts";
-import sodium from "./dist/browsers/sodium.js";
-
-export default sodium as Sodium;
-
-export * from "./basic_types.ts";
+// @deno-types="./libsodium_basic.d.ts"
+export * from "./dist/browsers/sodium.js";
+// @deno-types="./libsodium_basic.d.ts"
+import sodium from "./dist/browsers/sodium.js"
+export default sodium;
