+++ libsodium_common.ts
@@ -1,0 +1,8 @@
+// deno-lint-ignore-file camelcase
+
+export enum base64_variants {
+    ORIGINAL,
+    ORIGINAL_NO_PADDING,
+    URLSAFE,
+    URLSAFE_NO_PADDING,
+}
