+++ sumo.ts
@@ -1,8 +1,5 @@
-// Copyright 2020-present the denosaurs team. All rights reserved. MIT license.
-
-import { Sodium, SumoAddons } from "./sumo_types.ts";
-import sodium from "./dist/browsers-sumo/sodium.js";
-
-export default sodium as Sodium & SumoAddons;
-
-export * from "./sumo_types.ts";
+// @deno-types="./libsodium_sumo.d.ts"
+export * from "./dist/browsers-sumo/sodium.js";
+// @deno-types="./libsodium_sumo.d.ts"
+import sodium from "./dist/browsers-sumo/sodium.js"
+export default sodium;
